% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fns.R
\name{headtail}
\alias{headtail}
\title{Displays head and tail of an input df}
\usage{
headtail(df, n_head, n_tail)
}
\arguments{
\item{df}{Input data frame.}

\item{n_head}{Number of head to show.}

\item{n_tail}{Number of tail to show.}
}
\value{
A data frame showing first n_head and last n_tail rows of df.
}
\description{
Displays head and tail of an input df
}
\examples{
headtail(mtcars, 5, 5)
}
